api/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
joblib==1.2.0
python-dotenv==1.0.0
numpy==1.26.2
boto3==1.26.137
google-api-python-client==2.86.0
pandas==2.1.3
pyarabic==0.6.15
tashaphyne==0.3.6
emoji==2.2.0
python-multipart==0.0.6
scikit-learn==1.3.2